Fix config_flow using a HA selector that has never existed; add mypy, PARALLEL_UPDATES
The biggest find: config_flow.py called selector.SerialPortSelector() for
both the serial-setup and reconfigure steps. That class does not exist
anywhere in Home Assistant core - checked every *Selector class across
homeassistant.helpers.selector on the current release. Both steps would
raise AttributeError building their form schema, meaning fresh serial
setup and reconfiguring an existing serial entry were both broken.
Replaced with the actual standard pattern: enumerate ports via
serial.tools.list_ports.comports() and offer them through a
SelectSelector (custom_value=True so an undetected port can still be
typed by hand) - which also directly serves the original ask to probe
available ports and suggest the right one during setup.

Brought the whole package to a clean `mypy custom_components/davis_vantage/`
(0 errors, warn_unused_ignores on) and added it to CI:
- client.py: _vantagepro2 was typed as `None` (a bare `# type: ignore`
  suppressing the real type), so mypy couldn't verify any of its ~15 call
  sites. Properly typed it as `VantagePro2 | None` and added the same
  lazy-connect guard already used elsewhere to methods that were missing
  it (get_rain_collector, set_rain_collector,
  get_latitude_longitude_elevation, get_davis_time, set_davis_time,
  get_info, get_static_info, set_yearly_rain, set_archive_period).
- utils.get_wind_rose: widened int -> float (every real caller passes a
  float; the narrow hint was just wrong).
- sensor.py: found ANOTHER duplicate-key bug in the process - two
  DavisSensorEntityDescription entries both used key="ForecastIcon",
  meaning they'd generate the same unique_id and HA would silently drop
  whichever lost that race, so the disabled-by-default one (lowercase
  HA-vocabulary condition strings) could never actually be enabled. Gave
  it a distinct key. Also extracted the repeated
  `float(data.get(x)) if data.get(x) is not None else None` pattern (and
  the WindDir/Barometer range-check variants) into small named helpers -
  fixes the mypy narrowing issue and de-duplicates ~7 near-identical
  lambdas at once.
- Removed five now-dead `# type: ignore` comments (config_flow.py,
  services.py) that were suppressing errors that no longer exist.

Also added PARALLEL_UPDATES = 0 to all three platforms (sensor,
binary_sensor, weather) - Silver-tier requirement, correct here since
every entity reads from the already-polled coordinator rather than doing
its own I/O.

Added tests/test_sensor.py (24 tests) for the newly-extracted helpers.
Suite is now 111 tests, still green.

+++ custom_components/davis_vantage/utils.py
@@ -127,7 +127,7 @@ def calc_feels_like(temperature_f: float, humidity: float, windspeed_mph: float)
 def convert_to_iso_datetime(value: datetime, tzinfo: ZoneInfo) -> datetime:
     return value.replace(tzinfo=tzinfo)
 
-def get_wind_rose(bearing: int) -> str:
+def get_wind_rose(bearing: float) -> str:
     directions = ["N", "NNE", "NE", "ENE", "E", "ESE", "SE", "SSE", "S", "SSW", "SW", "WSW", "W", "WNW", "NW", "NNW"]
     index = round(bearing / 22.5) % 16
     return directions[index]
